Changing password quoting from simple to double quotes
Even though the [official documentation](https://dev.mysql.com/doc/refman/5.5/en/option-files.html)
states that both single and double-quotes are allowed for use in quoting the `password` option,
the `mysqlreport` tool (which is packaged in mysql-client-5-5 on Debian and Ubuntu) only matches
against double-quotes. Hence it was impossible to use this tool with the puppet module.

Signed-off-by: Mickaël FALCK <lastmikoi@lastmikoi.net>